+++ src/config.lua
@@ -2,33 +2,29 @@
 return {
 	enabled = true,
 	Artemis = {
-		ArtemisNoRequirements = true,
-		ArtemisSpawnAmount = 10,
+		enabled = true,
+		requirements = false,
 	},
-
 	Athena = {
-		AthenaNoRequirements = true,
-		AthenaSpawnAmount = 10,
+		enabled = true,
+		requirements = false,
 	},
-
 	Dionysus = {
-		DionysusNoRequirements = true,
-		DionysusSpawnAmount = 10,
+		enabled = true,
+		requirements = false,
 	},
 }, {
 	enabled = "Set to true to enable the mod, false to disable.",
 	Artemis = {
-		ArtemisNoRequirements = "Set to true to allow Artemis to spawn without restrictions.",
-		ArtemisSpawnAmount = "Set value to limit how many times Artemis can spawn throughout a run (If you have enabled requirements).",
+		enabled = "Set to true to enable Artemis, false to disable.",
+		requirements = "Set to true to force Artemis to spawn with restrictions, like Hermes does - and does not add to the MetGod list.",
 	},
-
 	Athena = {
-		AthenaNoRequirements = "Set to true to allow Athena to spawn without restrictions.",
-		AthenaSpawnAmount = "Set value to limit how many times Athena can spawn throughout a run (If you have enabled requirements).",
+		enabled = "Set to true to enable Athena, false to disable.",
+		requirements = "Set to true to force Athena to spawn with restrictions, like Hermes does - and does not add to the MetGod list.",
 	},
-
 	Dionysus = {
-		DionysusNoRequirements = "Set to true to allow Dionysus to spawn without restrictions.",
-		DionysusSpawnAmount = "Set value to limit how many times Dionysus can spawn throughout a run (If you have enabled requirements).",
+		enabled = "Set to true to enable Dionysus, false to disable.",
+		requirements = "Set to true to force Dionysus to spawn with restrictions, like Hermes does - and does not add to the MetGod list.",
 	},
 }