+++ CHANGELOG.md
@@ -1,7 +1,9 @@
 # Changelog
 
 ## [Unreleased]
-Fixed issue which wouldnt allow any traits in slots to overwrite already slotted boons.
+- Fixed issue which wouldnt allow any traits in slots to overwrite already slotted boons.
+- Reworked Config
+- DEV: Changed to function with GodsAPI
 
 ## [2.0.1] - 2025-10-24